Create read only versions of step 1 and 2 of the deploy modal
