+++ js/containers/DeployPlanRO.jsx
@@ -1,0 +1,33 @@
+const React = require('react');
+const ReactRedux = require('react-redux');
+
+const actions = require('../_actions.js');
+const deployment = require('../constants/deployment.js');
+
+function DeployPlanRO(props) {
+	return (
+		<div>
+			<div className="section deploy-plan">
+				<header id="1">Deployment Plan</header>
+				<div className="summary_title">{props.title}</div>
+				<p className="summary_of_changes">
+					{props.summary_of_changes}
+				</p>
+			</div>
+		</div>
+	);
+}
+
+const mapStateToProps = function(state) {
+	return {
+		title: state.plan.title,
+		summary_of_changes: state.plan.summary_of_changes
+	};
+};
+
+const mapDispatchToProps = function(dispatch) {
+	return {};
+};
+
+module.exports = ReactRedux.connect(mapStateToProps, mapDispatchToProps)(DeployPlanRO);
+
