+++ sass/components/deploy-plan.sass
@@ -1,0 +1,9 @@
+.deploy-plan
+
+	.summary_title
+		margin: 10px 0 20px
+		font-size: 16px
+		font-weight: bold
+
+	.summary_of_changes
+		white-space: pre-wrap
