+++ sass/components/target-release.sass
@@ -1,0 +1,5 @@
+@import "../_bootstrap-variables.sass"
+
+.target-release
+	table
+		margin-top: 10px
