+++ sass/style.sass
@@ -21,8 +21,12 @@ $textarea-transition-time: 0.6s
 @import "selectize/selectize"
 @import "selectize/selectize.bootstrap3"
 
+// react components
 @import 'components/current-build-status'
+@import 'components/deploy-plan'
 @import 'components/status-box'
+@import 'components/target-release'
+
 ul.list,
 ol.list
 	list-style: none
@@ -607,18 +611,22 @@ button.sidebar-open
 		font-size: 40px
 		content: "\2022"
 
-	&.status-Aborting,
-	&.status-Pending,
+	&.status-Aborting
+		color: $status-aborting
+	&.status-Pending
+		color: $status-pending
 	&.status-Submitted
-		color: $brand-warning
+		color: $status-submitted
 	&.status-Approved
-		color: $brand-success
+		color: $status-approved
 	&.status-Deploying
-		color: $brand-primary
-	&.status-Failed,
-	&.status-Invalid,
+		color: $status-deploying
+	&.status-Failed
+		color: $status-failed
+	&.status-Invalid
+		color: $status-invalid
 	&.status-Rejected
-		color: $brand-danger
+		color: $status-rejected
 
 	.deploy-status-text
 		padding: 3px 0 0 25px
