+++ static/components/deploy-plan.css
@@ -1,0 +1,7 @@
+.deploy-plan .summary_title {
+  margin: 10px 0 20px;
+  font-size: 16px;
+  font-weight: bold; }
+
+.deploy-plan .summary_of_changes {
+  white-space: pre-wrap; }
