+++ static/components/target-release.css
@@ -1,0 +1,2 @@
+.target-release table {
+  margin-top: 10px; }
